+++ addOns/reports/src/test/java/org/zaproxy/addon/reports/ExtensionReportsUnitTest.java
@@ -35,7 +35,11 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
+import java.util.List;
 import java.util.Locale;
+import javax.xml.parsers.ParserConfigurationException;
+import org.apache.commons.httpclient.URI;
+import org.apache.commons.httpclient.URIException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -45,6 +49,9 @@ import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.extension.ExtensionLoader;
 import org.parosproxy.paros.model.Model;
+import org.parosproxy.paros.network.HttpMalformedHeaderException;
+import org.parosproxy.paros.network.HttpMessage;
+import org.xml.sax.SAXException;
 import org.zaproxy.addon.automation.jobs.PassiveScanJobResultData;
 import org.zaproxy.zap.extension.alert.AlertNode;
 import org.zaproxy.zap.extension.pscan.PluginPassiveScanner;
@@ -301,7 +308,7 @@ public class ExtensionReportsUnitTest {
         ReportData reportData = new ReportData();
         AlertNode root = new AlertNode(0, "Test");
         reportData.setAlertTreeRootNode(root);
-        reportData.setSites(Arrays.asList("test"));
+        reportData.setSites(Arrays.asList("http://example.com"));
         return reportData;
     }
 
@@ -448,5 +455,145 @@ public class ExtensionReportsUnitTest {
         assertThat(report.contains(HTML_REPORT_INSTANCE_SUMMARY_SECTION_TITLE), is(false));
         assertThat(report.contains(HTML_REPORT_ALERT_DETAILS_SECTION_TITLE), is(false));
         assertThat(report.contains(HTML_REPORT_PASSING_RULES_SECTION_TITLE), is(true));
+    }
+
+    private static AlertNode getAlertNode(String name, String desc, int risk, int confidence)
+            throws URIException, HttpMalformedHeaderException {
+        AlertNode node = new AlertNode(risk, name);
+        Alert alert = new Alert(1, risk, confidence, name);
+        String uriStr = "http://example.com/example_" + risk;
+
+        HttpMessage msg = new HttpMessage(new URI(uriStr, true));
+        msg.setRequestBody("Test Request Body");
+        msg.setResponseBody("Test Response Body");
+
+        alert.setDetail(
+                desc,
+                uriStr,
+                "Test Param",
+                "Test Attack",
+                "Test Other",
+                "Test Solution",
+                "Test Reference",
+                "Test Evidence",
+                123,
+                456,
+                msg);
+        node.setUserObject(alert);
+
+        AlertNode instance = new AlertNode(0, name);
+        instance.setUserObject(alert);
+
+        node.add(instance);
+
+        return node;
+    }
+
+    private static ReportData getTestReportDataWithAlerts()
+            throws URIException, HttpMalformedHeaderException {
+        ReportData reportData = new ReportData();
+        reportData.setTitle("Test Title");
+        reportData.setDescription("Test Description");
+        reportData.setIncludeAllConfidences(true);
+        reportData.setIncludeAllRisks(true);
+        List<PluginPassiveScanner> list = new ArrayList<PluginPassiveScanner>();
+        PassiveScanJobResultData pscanData = new PassiveScanJobResultData("passiveScan-wait", list);
+        reportData.addReportObjects(pscanData.getKey(), pscanData);
+
+        AlertNode root = new AlertNode(0, "Test");
+        reportData.setAlertTreeRootNode(root);
+        root.add(getAlertNode("XSS", "XSS Description", Alert.RISK_HIGH, Alert.CONFIDENCE_MEDIUM));
+        reportData.setSites(Arrays.asList("http://example.com"));
+        return reportData;
+    }
+
+    private static File generateReportWithAlerts(Template template, File f)
+            throws IOException, DocumentException {
+        ExtensionReports extRep = new ExtensionReports();
+        ReportData reportData = getTestReportDataWithAlerts();
+        reportData.setSections(template.getSections());
+        return extRep.generateReport(reportData, template, f.getAbsolutePath(), false);
+    }
+
+    private static String cleanReport(String str) {
+        return str.replaceFirst("generated=\".*\"", "generated=\"DATE\"")
+                .replaceAll("basic-.*/", "dir")
+                .replaceAll("[\\n\\r\\t]", "");
+    }
+
+    @ParameterizedTest
+    @ValueSource(
+            strings = {
+                "traditional-md",
+                "traditional-xml",
+                "traditional-html",
+                "traditional-html-plus"
+            })
+    public void shouldGenerateExpectedReport(String templateName)
+            throws IOException, DocumentException, SAXException, ParserConfigurationException {
+        // Given
+        File t = new File("src/main/zapHomeFiles/reports/" + templateName + "/template.yaml");
+        Template template = new Template(t);
+        String fileName = "basic-" + templateName;
+        File f = File.createTempFile(fileName, template.getExtension());
+
+        // When
+        File r = generateReportWithAlerts(template, f);
+        String report = new String(Files.readAllBytes(r.toPath()));
+
+        File expectedReport =
+                new File(
+                        "src/test/resources/org/zaproxy/addon/automation/resources/"
+                                + fileName
+                                + "."
+                                + template.getExtension());
+        String expected = new String(Files.readAllBytes(expectedReport.toPath()));
+
+        // Then
+        assertThat(cleanReport(report), is(equalTo(cleanReport(expected))));
+    }
+
+    private static void generateTestFile(String templateName)
+            throws IOException, DocumentException {
+
+        Template template =
+                new Template(
+                        new File(
+                                "src/main/zapHomeFiles/reports/"
+                                        + templateName
+                                        + "/template.yaml"));
+        generateReportWithAlerts(
+                template,
+                new File(
+                        "src/test/resources/org/zaproxy/addon/automation/resources/basic-"
+                                + templateName
+                                + "."
+                                + template.getExtension()));
+    }
+
+    /**
+     * This can be used to either regenerate the test files when they are expected to have changed
+     * so that they can be checked in or to make it easier to see the differences if the tests fail.
+     *
+     * @param args not used
+     */
+    public static void main(String[] args) {
+        try {
+            Constant.messages = new I18N(Locale.ENGLISH);
+
+            Model model = mock(Model.class, withSettings().defaultAnswer(CALLS_REAL_METHODS));
+            Model.setSingletonForTesting(model);
+            ExtensionLoader extensionLoader = mock(ExtensionLoader.class, withSettings().lenient());
+            Control.initSingletonForTesting(Model.getSingleton(), extensionLoader);
+            Model.getSingleton().getOptionsParam().load(new ZapXmlConfiguration());
+
+            generateTestFile("traditional-md");
+            generateTestFile("traditional-xml");
+            generateTestFile("traditional-html");
+            generateTestFile("traditional-html-plus");
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 }