+++ addOns/reports/CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-### Changed
+
+### Fixed
 - Correct logging of dependency.
+- Inconsistencies between traditional reports and the 'old' core ones
 
 ### Fixed
 - Do not rely on default encoding when creating the reports, use UTF-8 always (Issue 6561).